code/requirements.txt
numpy
torch
scikit-learn
optuna
mysql
plotly
kaleido
tqdm
pandas
open3d
opencv-python
torchvision

Pillow
matplotlib
PyYAML